+++ src/fields/ExternalIdentity.php
@@ -1,9 +1,5 @@
 <?php
 
-/**
- * @copyright  Copyright (c) Flipbox Digital Limited
- */
-
 namespace flipbox\saml\sp\fields;
 
 use craft\base\ElementInterface;
@@ -17,28 +13,4 @@ class ExternalIdentity extends AbstractExternalIdentity
 
     use SamlPluginEnsured;
 
-    /**
-     * @param $value
-     * @param ElementInterface $element
-     * @return string
-     * @throws \Twig_Error_Loader
-     * @throws \yii\base\Exception
-     */
-    public function getStaticHtml($value, ElementInterface $element): string
-    {
-        if (! ($value instanceof Query)) {
-            return '';
-        }
-
-        return \Craft::$app->getView()->renderTemplate(
-            'saml-sp/_cp/fields/external-id',
-            [
-                'identities' => $value,
-                'element'    => $element,
-                'baseProviderUrl' => UrlHelper::cpUrl(
-                    'saml-sp/metadata'
-                ),
-            ]
-        );
-    }
 }