+++ src/services/messages/Metadata.php
@@ -1,10 +1,4 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: dsmrt
- * Date: 1/9/18
- * Time: 9:48 AM
- */
 
 namespace flipbox\saml\sp\services\messages;
 
@@ -16,21 +10,6 @@ class Metadata extends AbstractMetadata implements MetadataServiceInterface
 {
     use SamlPluginEnsured;
 
-    /**
-     * Deprecated Constants
-     */
-    /**
-     * @deprecated
-     */
-    const LOGIN_LOCATION = 'sso/login';
-    /**
-     * @deprecated
-     */
-    const LOGOUT_RESPONSE_LOCATION = 'sso/logout';
-    /**
-     * @deprecated
-     */
-    const LOGOUT_REQUEST_LOCATION = 'sso/logout/request';
     /**
      * @return array
      */